pyproject.toml
[project]
name = "csu-2026-capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "ray[default]>=2.54.1",
    "sqlalchemy>=2.0",
    "pillow>=10.0",
]
